Support 'read'.
The issue we were encountering was that we were looking up `gets` in the GOT,
however, since `gets` hadn't been called yet, it hadn't been resolved yet,
thus we got a bad resolved address for `gets`. Since we're dumping the
address via `puts` atm, we know that `puts` will be present.

This is another thing we'll want to add as a check (asserting that the
leaked address has been GOT resolved).