Add initial support for pytest.
nose2 is old and no longer maintained, switching to pytest.

At the moment, we still use a few trivial nose2 functions like ok_.

These should be easily transformable into asserts.

Bigger change is that pytest captures logs as well as stdout,
but this can be disabled. Unfortunately, it seems that the obvious
way of disabling it also pulls in the very noisy logs from angr
as well.